Spacetime: np->torch
Additionally, one assertion in fermionMatrix is a shape comparison and
requires spacetime.dims be a shape.

That comparison doesn't need a .all() reduction as long as spacetime.dims
is a shape.  Get the shape from .T, but could equivalently get the shape
from .X or .Y

In the example we need to make nt a (trivial) tensor