No explicit casts (#24)
Closes #13.

* Convert to complex by +0j rather than .to

This saves us from having to specify a type like complex64 or complex128;
it just promotes to the complex type corresponding to the float we've
got.

* Use +0j type promotion trick in examples.

* _demos work; trouble-making float32s eliminated!

Something annoying that took a long time to track down was that in the definition of some _demo functions some parameters got torch.tensor default values whose dtype was set at import time, which can be before a user might torch.set_default_dtype(torch.float64) (for example). Since the default dtype is float32 you can run into type conflicts. Extremely annoying.

The way around this was to default them to None and then check for None and create them. Annoying, but safe.

* Some casts that remain are:

    - casts to integers
    - casts where we know another tensor of the type we need but don't have to explicitly specify the precision