Regression fix: Canonical projection succeeds
sanity-checks/canonical-double-occupancy.py was failing, but there was
not a numerical issue; evaluating _canonical_weights failed.

The issue was that as part of cleaning up the UU observables, the fermion
determinant was not made @observable.  This was a bug introduced in commit
d6d6cd4d4744b2b2055bf11199ac5fdce9ad603d
and somehow I hadn't run the since.

This returns the @observable decorator to _detUUPlusOne, which is needed
for the _canonical_weights.  As soon as it was restored the sanity check
passed.