Bootstrapping can compute ensemble intermediates (#100)
I tried running

docs/sanity-checks/free-theory-{continuum,infinite-volume}-limit.ipynb

and found that I couldn't compute pairing observables because the
bootstrap no longer had access to G.

G was marked @intermediate in #84, but in my trial production I was
manually storing the propagator and therefore wasn't noticing this
issue.

Intermediate quantities (typically) are not needed at the bootstrap
level, but the propagator is an exception if you try to compute pairing
matrix observables.  There's no reason I can think of, though, to make
it *disallowed* for the bootstrap to compute the intermediate if it
knows what it's doing.